\name{bedGraphCoverage}
\alias{bedGraphCoverage}
\title{bedGraphCoverage}
\description{Compute total coverage from a bedGraph file.}
\usage{bedGraphCoverage(input.bedGraph)}
\arguments{
  \item{input.bedGraph}{Input bedGraph file.}
}

\value{data.table of file name and total coverage.}

\author{Toby Dylan Hocking}





