\name{bigWigCoverage}
\alias{bigWigCoverage}
\title{bigWigCoverage}
\description{Compute total and average coverage from a bigWig file.}
\usage{bigWigCoverage(input.bigWig)}
\arguments{
  \item{input.bigWig}{Input bigWig file.}
}

\value{one row data.table with columns total.bases, mean.coverage,
total.coverage, and input.bedGraph.}

\author{Toby Dylan Hocking}





