\name{bigWigToBedGraph}
\alias{bigWigToBedGraph}
\title{bigWigToBedGraph}
\description{Run bigWigToBedGraph command line program.}
\usage{bigWigToBedGraph(in.bigWig, 
    out.bedGraph, chrom = NULL, 
    start = NULL, end = NULL)}
\arguments{
  \item{in.bigWig}{character string path to input bigWig file.}
  \item{out.bedGraph}{character string path to output bedGraph file.}
  \item{chrom}{character string, chromosome name to filter data.}
  \item{start}{\code{start} position to filter data.}
  \item{end}{\code{end} position to filter data.}
}



\author{Toby Dylan Hocking}





