\name{convert_labels}
\alias{convert_labels}
\title{convert labels}
\description{Convert label text files in proj.dir/labels/*.txt to
proj.dir/samples/*/*/labels.bed files.}
\usage{convert_labels(proj.dir, verbose = 0)}
\arguments{
  \item{proj.dir}{project directory.}
  \item{verbose}{Print messages?}
}



\author{Toby Dylan Hocking}





