\name{create_problems_all}
\alias{create_problems_all}
\title{create problems all}
\description{Create problems directory with sub-directories for each problem
with labels.}
\usage{create_problems_all(data.dir.arg, PBS.header = getOption("PeakSegPipeline.header", 
    PBS.header.default))}
\arguments{
  \item{data.dir.arg}{a directory with labels.bed and coverage.bedGraph.}
  \item{PBS.header}{Header for sh files.}
}



\author{Toby Dylan Hocking}





