\name{downloadProblems}
\alias{downloadProblems}
\title{downloadProblems}
\description{Create problems.bed based on data from UCSC.}
\usage{downloadProblems(genome, problems.bed, url.prefix = "http://hgdownload.soe.ucsc.edu/goldenPath/")}
\arguments{
  \item{genome}{UCSC ID e.g. hg19 or hg38.}
  \item{problems.bed}{file to save.}
  \item{url.prefix}{http://path.to/goldenPath/}
}



\author{Toby Dylan Hocking}





