\name{gap2problems}
\alias{gap2problems}
\title{gap2problems}

\usage{gap2problems(gap.bed, chromInfo.txt, problems.bed)}
\arguments{
  \item{gap.bed}{gap.bed file, parts of the genome with gaps (NNNN).}
  \item{chromInfo.txt}{chromInfo.txt file, chromosome sizes.}
  \item{problems.bed}{will be created}
}



\author{Toby Dylan Hocking}





