\name{pasteQuote}
\alias{pasteQuote}
\title{pasteQuote}
\description{Paste arguments together after quoting.}
\usage{pasteQuote(...)}
\arguments{
  \item{\dots}{Arguments to be pasted together.}
}

\value{Character string suitable for passing to \code{\link[base]{system}}.}

\author{Toby Dylan Hocking}





