\name{pipeline}
\alias{pipeline}
\title{pipeline}
\description{Run entire PeakSegFPOP + PeakSegJoint pipeline.}
\usage{pipeline(set.dir.path, verbose = 0)}
\arguments{
  \item{set.dir.path}{data set directory.}
  \item{verbose}{print messages?}
}



\author{Toby Dylan Hocking}





