\name{plot_all}
\alias{plot_all}
\title{plot all}
\description{Plot results of peak calling, generate summary web page.}
\usage{plot_all(set.dir.arg, zoom.out.times = 10)}
\arguments{
  \item{set.dir.arg}{Path/to/data/dir.}
  \item{zoom.out.times}{
}
}



\author{Toby Dylan Hocking}





