\name{problem.features}
\alias{problem.features}
\title{problem features}
\description{Compute features for one segmentation problem.}
\usage{problem.features(problem.dir)}
\arguments{
  \item{problem.dir}{problemID directory with problemID/coverage.bedGraph.}
}

\value{Nothing, but writes problemID/features.tsv if it does not exist
already.}

\author{Toby Dylan Hocking}





