\name{problem.joint.plot}
\alias{problem.joint.plot}
\title{problem joint plot}
\description{Plot one chunk.}
\usage{problem.joint.plot(chunk.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{chunk.dir}{project/problems/problemID/chunks/chunkID}
  \item{verbose}{
}
}

\value{Nothing}

\author{Toby Dylan Hocking}





