\name{problem.joint}
\alias{problem.joint}
\title{problem joint}
\description{Fit a joint model.}
\usage{problem.joint(jointProblem.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{jointProblem.dir}{path/to/jointProblem}
  \item{verbose}{
}
}

\value{Model from PeakSegJointFaster.}

\author{Toby Dylan Hocking}





