\name{problem.joint.target}
\alias{problem.joint.target}
\title{problem joint target}
\description{Compute target interval for a joint problem.}
\usage{problem.joint.target(jointProblem.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{jointProblem.dir}{Joint problem directory.}
  \item{verbose}{
}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





