\name{problem.joint.train}
\alias{problem.joint.train}
\title{problem joint train}
\description{Learn a penalty function for joint peak prediction.}
\usage{problem.joint.train(data.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{data.dir}{project directory.}
  \item{verbose}{
}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





