\name{problem.predict}
\alias{problem.predict}
\title{problem predict}
\description{Predict peaks for a genomic segmentation problem.}
\usage{problem.predict(problem.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{problem.dir}{project/samples/groupID/sampleID/problems/problemID.}
  \item{verbose}{
}
}

\value{data.table of peak predictions.}

\author{Toby Dylan Hocking}





