\name{problem.table}
\alias{problem.table}
\title{problem table}
\description{Convert a path with a problem string to a data.table.}
\usage{problem.table(problem.dir)}
\arguments{
  \item{problem.dir}{path with a problem string, e.g. chrX:6000-1000000}
}

\value{data.table with columns chrom, problemStart, problemEnd.}

\author{Toby Dylan Hocking}





