\name{problem.tempfile}
\alias{problem.tempfile}
\title{problem tempfile}
\description{Create a (problem,penalty)-specific temporary file name to pass to
PeakSegFPOP_dir as the cost function database.}
\usage{problem.tempfile(problem.dir, 
    pen.str)}
\arguments{
  \item{problem.dir}{full path to problem directory.}
  \item{pen.str}{penalty string.}
}

\value{character: temporary file name.}

\author{Toby Dylan Hocking}





