\name{problem.train}
\alias{problem.train}
\title{problem train}
\description{Train a penalty function that predicts the number of peaks for
each separate problem. Run this step after computing target
intervals for each labeled separate problem (\code{\link{problem.target}}), and
before peak prediction (\code{\link{problem.predict}}).}
\usage{problem.train(data.dir.str, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{data.dir.str}{
}
  \item{verbose}{
}
}



\author{Toby Dylan Hocking}





