\name{psp_lapply}
\alias{psp_lapply}
\title{psp lapply}
\description{Use the lapply-like function defined in the PeakSegPipeline.lapply
option, or the default \code{\link[base]{lapply}}.}
\usage{psp_lapply(...)}
\arguments{
  \item{\dots}{Passed to the lapply-like function.}
}

\value{Result of the lapply-like function.}

\author{Toby Dylan Hocking}





