\name{readBigWigSamples}
\alias{readBigWigSamples}
\title{readBigWigSamples}
\description{read a bunch of bigwig files into R as a list of data.frames that
can be passed to PeakSegJointSeveral.}
\usage{readBigWigSamples(problem, bigwig.file.vec)}
\arguments{
  \item{problem}{
}
  \item{bigwig.file.vec}{
}
}



\author{Toby Dylan Hocking}





