\name{Rscript}
\alias{Rscript}
\title{Rscript}
\description{Run sprintf and paste Rscript -e before -- useful for constructing
command lines.}
\usage{Rscript(...)}
\arguments{
  \item{\dots}{Passed to sprintf -- there should be no single quotes.}
}

\value{Command line.}

\author{Toby Dylan Hocking}





