\name{stop.without.ucsc}
\alias{stop.without.ucsc}
\title{stop without ucsc}
\description{Stop with an error if UCSC command line programs are not
available.}
\usage{stop.without.ucsc(prog.vec = c("bigWigInfo", "bigWigToBedGraph", 
    "bedToBigBed", "bedGraphToBigWig"))}
\arguments{
  \item{prog.vec}{
}
}

\value{Full paths to programs.}

\author{Toby Dylan Hocking}





