\name{create_problems_all}
\alias{create_problems_all}
\title{create problems all}

\usage{create_problems_all(data.dir.arg, PBS.header = "#!/bin/bash\\n#PBS -l nodes=1:ppn=4\\n#PBS -l walltime=24:00:00\\n#PBS -A bws-221-ae\\n#PBS -m ae\\n#PBS -M tdhock5@gmail.com\\n#PBS -V")}
\arguments{
  \item{data.dir.arg}{
}
  \item{PBS.header}{
}
}



\author{Toby Dylan Hocking}





