\name{create_problems_joint}
\alias{create_problems_joint}
\title{create problems joint}
\description{Create joint problems for one separate problem, after separate
peak prediction.}
\usage{create_problems_joint(prob.dir)}
\arguments{
  \item{prob.dir}{proj.dir/problems/problemID}
}



\author{Toby Dylan Hocking}





