\name{create_track_hub}
\alias{create_track_hub}
\title{create track hub}

\usage{create_track_hub(data.dir = "test/input", url.prefix = "http://hubs.hpc.mcgill.ca/~thocking/PeakSegFPOP-", 
    genome = "hg19", email = "email@domain.com", goldenPath.url = "http://hgdownload.soe.ucsc.edu/goldenPath/")}
\arguments{
  \item{data.dir}{
}
  \item{url.prefix}{
}
  \item{genome}{
}
  \item{email}{
}
  \item{goldenPath.url}{
}
}



\author{Toby Dylan Hocking}





