\name{jobs_create_run}
\alias{jobs_create_run}
\title{jobs create run}
\description{Run entire PeakSegFPOP + PeakSegJoint pipeline.}
\usage{jobs_create_run(set.dir.path, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{set.dir.path}{data set directory.}
  \item{verbose}{print messages?}
}



\author{Toby Dylan Hocking}





