\name{jobs_create_run}
\alias{jobs_create_run}
\title{jobs create run}
\description{Run entire PeakSegFPOP + PeakSegJoint pipeline.}
\usage{jobs_create_run(set.dir.path, 
    ...)}
\arguments{
  \item{set.dir.path}{data set directory.}
  \item{\dots}{passed to \code{\link{jobs_create}}}
}



\author{Toby Dylan Hocking}





