\name{problem.joint.predict.job}
\alias{problem.joint.predict.job}
\title{problem joint predict job}
\description{Compute all joint peak predictions for the joint problems listed
in jobProblems.bed}
\usage{problem.joint.predict.job(job.dir)}
\arguments{
  \item{job.dir}{project/jobs/jobID}
}

\value{data.table of predicted loss, peak positions, and means per sample
(in a list column).}

\author{Toby Dylan Hocking}





