\name{problem.joint.predict}
\alias{problem.joint.predict}
\title{problem joint predict}
\description{Compute peak predictions for a joint problem.}
\usage{problem.joint.predict(jointProblem.dir)}
\arguments{
  \item{jointProblem.dir}{project/problems/problemID/jointProblems/jointProbID}
}

\value{list of peaks and loss.}

\author{Toby Dylan Hocking}





