\name{problem.joint.target}
\alias{problem.joint.target}
\title{problem joint target}
\description{Compute target interval for a joint problem.}
\usage{problem.joint.target(jointProblem.dir)}
\arguments{
  \item{jointProblem.dir}{Joint problem directory.}
}

\value{list of output from PeakSegJointError.}

\author{Toby Dylan Hocking}





