\name{problem.joint.targets}
\alias{problem.joint.targets}
\title{problem joint targets}
\description{Compute targets for a separate problem.}
\usage{problem.joint.targets(problem.dir)}
\arguments{
  \item{problem.dir}{project/problems/problemID}
}

\value{Named list of two matrices: targets and features.}

\author{Toby Dylan Hocking}





