\name{problem.joint.targets}
\alias{problem.joint.targets}
\title{problem joint targets}
\description{Compute targets for a separate problem.}
\usage{problem.joint.targets(problem.dir)}
\arguments{
  \item{problem.dir}{project/problems/problemID}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





