\name{problem.joint.targets.train}
\alias{problem.joint.targets.train}
\title{problem joint targets train}
\description{Compute all joint target intervals then learn joint penalty
functions, in parallel over problems using
\code{\link[future.apply]{future_lapply}}.}
\usage{problem.joint.targets.train(data.dir, 
    verbose = getOption("PeakSegPipeline.verbose", 
        1))}
\arguments{
  \item{data.dir}{project directory.}
  \item{verbose}{
}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





