\name{problem.joint.targets.train}
\alias{problem.joint.targets.train}
\title{problem joint targets train}
\description{Compute all target intervals then learn a penalty function.}
\usage{problem.joint.targets.train(data.dir)}
\arguments{
  \item{data.dir}{project directory.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





