\name{problem.joint.train}
\alias{problem.joint.train}
\title{problem joint train}
\description{Learn a penalty function for joint peak prediction.}
\usage{problem.joint.train(data.dir)}
\arguments{
  \item{data.dir}{project directory.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





