\name{problem.labels}
\alias{problem.labels}
\title{problem labels}
\description{read problemID/labels.bed if it exists, otherwise read
sampleID/labels.bed}
\usage{problem.labels(problem.dir)}
\arguments{
  \item{problem.dir}{
}
}

\value{data.table with one row for each label and columns chrom,
chromStart, chromEnd, annotation.}

\author{Toby Dylan Hocking}





