\name{problem.pred.cluster.targets}
\alias{problem.pred.cluster.targets}
\title{problem pred cluster targets}
\description{For a given problem, predict independently for each sample, then
cluster peaks across samples, then compute joint target intervals.}
\usage{problem.pred.cluster.targets(proj.dir)}
\arguments{
  \item{proj.dir}{project directory.}
}

\value{List of features and target matrices (same as \code{\link{problem.joint.target}}).}

\author{Toby Dylan Hocking}





