\name{problem.predict.allSamples}
\alias{problem.predict.allSamples}
\title{problem predict allSamples}
\description{Predict for all samples in parallel.}
\usage{problem.predict.allSamples(prob.dir)}
\arguments{
  \item{prob.dir}{project/problems/problemID directory.}
}

\value{data.table of predicted peaks.}

\author{Toby Dylan Hocking}





