\name{system.or.stop}
\alias{system.or.stop}
\title{system or stop}
\description{Run a command line or stop with an error.}
\usage{system.or.stop(cmd, verbose = 0)}
\arguments{
  \item{cmd}{Command line, passed to system.}
  \item{verbose}{print output?}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





